Support multiple package sources and copy concurrently

+++ src/Main.hs
@@ -4,7 +4,8 @@
 module Main where
 
 import           Control.Exception (throw)
-import           Control.Monad (when, forM_)
+import           Control.Monad (when, unless, forM)
+import           Control.Concurrent.Async (mapConcurrently)
 import           Control.Conditional (unlessM, ifM)
 import           Data.Aeson (encode, decode)
 import           Data.Aeson.TH
@@ -13,11 +14,13 @@ import qualified Data.ByteString.Lazy as Bz
 import           Data.Char (toLower)
 import           Data.Functor ((<$>))
 import           Data.Map (Map, toList)
-import           Data.Maybe (maybe)
+import           Data.Maybe (maybe, catMaybes)
 import           Data.Version (showVersion)
 import           Options.Applicative
 import           System.Directory (doesFileExist, doesDirectoryExist, removeDirectoryRecursive)
+import           System.Exit (exitFailure)
 import           System.FilePath ((</>), (<.>), normalise)
+import           System.IO (stderr, hPutStrLn)
 
 import           Paths_libget (version)
 
@@ -26,16 +29,19 @@ import           Utils (jsonField)
 
 
 data CmdOptions = CmdOptions
-  { _file     :: Maybe String,
-    _root     :: String,
-    _packages :: String
+  { _optPackageRoots :: [FilePath]
+  , _optFile         :: Maybe FilePath
+  , _optRoot         :: FilePath
   }
 
 data Dependency = Dependency
   { _depName        :: !String
   , _depVersion     :: !String
-  } deriving (Show, Eq)
+  } deriving (Eq)
 $(deriveJSON defaultOptions {fieldLabelModifier=jsonField} ''Dependency)
+
+instance Show Dependency where
+  show (Dependency name ver) = "dependency " ++ name ++ "@" ++ ver
 
 data Spec = Spec
   { _specVersion      :: !String
@@ -45,10 +51,6 @@ $(deriveJSON defaultOptions {fieldLabelModifier=jsonField} ''Spec)
 
 
 supportedSpecs = ["1"] :: [String]
-
-
-srcDir :: FilePath -> Dependency -> FilePath
-srcDir packageDir (Dependency name ver) = packageDir </> name </> ver
 
 
 copyDir' :: FilePath -> FilePath -> IO ()
@@ -77,27 +79,57 @@ rmdir :: FilePath -> IO ()
 rmdir dir = doesDirectoryExist dir >>= (`when` removeDirectoryRecursive dir)
 
 
-install:: FilePath -> FilePath -> Dependency -> IO ()
-install packageRoot dst dep =
-  unlessM (alreadyUpToDate dst dep) $ do
-    rmdir dst
-    copyDir' (srcDir packageRoot dep) dst
-    leaveCrumb dst dep
+first :: (Monad m, Functor m) => (a -> m (Maybe b)) -> [a] -> m (Maybe b)
+first f as = (safeHead . catMaybes) <$> mapM f as
+  where safeHead [] = Nothing
+        safeHead xs = Just (head xs)
+
+
+errorMsg :: String -> IO ()
+errorMsg msg = hPutStrLn stderr $ "Error: " ++ msg
+
+
+install :: [FilePath] -> FilePath -> Dependency -> IO Bool
+install packageRoots dst dep = do
+  done <- alreadyUpToDate dst dep
+  if done
+    then return True
+    else do
+      src' <- first existing (packageDir dep <$> packageRoots)
+      case src' of
+        Nothing -> do
+          errorMsg $ "failed to find package for " ++ show dep
+          return False
+        Just src -> do
+          rmdir dst
+          copyDir' src dst
+          leaveCrumb dst dep
+          return True
+  where
+    existing dir = ifM (doesDirectoryExist dir) (return $ Just dir) (return Nothing)
+    packageDir (Dependency name ver) packageRoot = packageRoot </> name </> ver
 
 
 main' :: CmdOptions -> IO ()
-main' (CmdOptions file root packageRoot) = do
+main' (CmdOptions packageRoots file root) = do
   spec' <- decode <$> content
   case spec' of
     Nothing   -> throw $ userError "did not understand specification input"
-    Just spec -> do
-      when (_specVersion spec `notElem` supportedSpecs) $
-        throw $ userError "input version is not supported"
-      forM_ (toList $ _specDependencies spec) $ \(dst, dep) ->
-        install packageRoot (root </> normalise dst) dep
+    Just spec -> handleSpec spec
   where
     content = maybe Bz.getContents Bz.readFile file
 
+    handleSpec spec = do
+      when (_specVersion spec `notElem` supportedSpecs) $
+        throw $ userError "input version is not supported"
+
+      results <- mapConcurrently installAtRoot (toList $ _specDependencies spec)
+      unless (and results) exitFailure
+
+    installAtRoot (dst, dep) = install packageRoots (root </> normalise dst) dep
+
+
+progName = "libget" :: String
 
 main :: IO ()
 main = execParser opts >>= main'
@@ -108,14 +140,15 @@ main = execParser opts >>= main'
 
 
 addVersion :: Parser (a -> a)
-addVersion = infoOption ("libget version " ++ showVersion version)
+addVersion = infoOption (progName ++ " version " ++ showVersion version)
   ( long "version"
-  <> help "Show version information" )
+  <> help "Show version information")
 
 
 cmdOptions :: Parser CmdOptions
 cmdOptions = CmdOptions
-  <$> (optional . strOption)
+  <$> some (argument str (metavar "PACKAGE_ROOTS..."))
+  <*> (optional . strOption)
       ( long    "file"
      <> short   'f'
      <> metavar "FILE"
@@ -127,8 +160,3 @@ cmdOptions = CmdOptions
      <> value   "."
      <> showDefault
      <> help    "Path to root of project" )
-  <*> strOption
-      ( long    "packages"
-     <> short   'p'
-     <> metavar "DIR"
-     <> help    "Root path to packages" )
